+++ ClassicLootManager/Modules/Common/Changelog/Changelog.lua
@@ -1,6 +1,26 @@
 local  _, CLM = ...
 CLM.ChangelogData = {
     {
+        version = "v4.3.0",
+        data = {
+            {
+                name = "Features",
+                data = {
+                    { header = "Item Tracker", body = "New sub-addon has been added (and enabled by default). This adds new auctioning column and tooltip information to each item, about who and how many has already. You can disable that in configuration or disable the addon at all." },
+                    { header = "Auction time increasing", body = "You can now enable automatic auction time increasing which will lineary multiply provided auction time by the number of items." },
+                    { header = "Trial of the Grand Crusader", body = "CLM is now ready for ToGC. Trophy and Regalia will now be classified as `Non-equippable` for default and dynamic item value classification. Their ilvl is now adjusted internally (245/258). ToGC bosses have now their respective encounter ids mapped since old ones are used same as for Ulduar." },
+                    { header = "Auction and Bidding UI slight refresh", body = "With trophies and regalia feeting  5 slots, up to 5 upgraded items will now be displayed when bidding on item. This required slight UI rework and resulted in `Pass` being moved and `Next` button added (to skip bidding on the current item without sending information to auctioneer)." },
+                },
+            },
+            {
+                name = "Fix",
+                data = {
+                    { header = "Tooltip", body = "Fixed timing issues with tooltip requests happening before CLM was ready." },
+                },
+            },
+        },
+    },
+    {
         version = "v4.2.3",
         data = {
             {